Moved function back to parent file. Reasoning: element is not used across site, only in script page. Leave it privated to script page.
