--- src/components/ScriptInputBox.css
@@ -1,118 +1,0 @@
-.input-box {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    > h2 {
-        margin: 10px 0px;
-    }
-}
-
-.input-container {
-    width: 50%;
-    min-width: 180px;
-    height: 50px;
-
-    background: var(--button-color);
-    color: white;
-
-    border-radius: 20px;
-    padding: 10px;
-    margin: 10px;
-
-    display: flex;
-    flex-direction: column;
-    align-items: start;
-    justify-content: center;
-
-    > div > b, p {
-        font-size: 12px;
-
-        color: #A6A9B4;
-        margin: 0px;
-    }
-}
-
-.test-input-box {
-    width: 85%;
-
-    font-size: 16px;
-    font-family: inherit;
-    font-weight: bold;
-
-    border-radius: 5px;
-}
-
-/* Styling for input boxes of different types*/
-input[type="text"] {
-    padding: 0px;
-    background: transparent;
-    border: none;
-    outline: none;
-    color: var(--dark-purple);
-}
-
-input[type="number"] {
-    padding: 0px;
-    background: transparent;
-    border: none;
-    outline: none;
-    color: var(--dark-purple);
-}
-input[type="number"]::-webkit-inner-spin-button,
-input[type="number"]::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-}
-
-/* Hide the browser's default checkbox */
-.checkbox-container {
-    width: 100%;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-    input {
-        position: absolute;
-        opacity: 0;
-        cursor: pointer;
-        height: 0;
-        width: 0;
-    }
-}
-
-/* Create a custom checkbox */
-.tf-container {
-    width: 100%;
-
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-}
-
-.boolean-checkbox {
-    width: 75px;
-    height: 20px;
-    margin: 0px 3px;
-
-    font-weight: bold;
-
-    border: 3px solid var(--dark-purple);
-    border-radius: 20px;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    cursor: pointer;
-}
-.boolean-checkbox.checked {
-    background: var(--dark-purple);
-    color: var(--button-color);
-}
-.boolean-checkbox:not(.checked) {
-    background: var(--button-color);
-    color: var(--dark-purple);
-}
